requirements.txt
websocket_client==1.4.2
